Bugfix/12323/fix-wrong-status-for-history-groups (#845)
* change calculations for hasMissingEvents

* lint fix

* Make time select in date filter editable (#838)

The datepicker used allows users to select time range. We populate the time select with time interval of 15 minutes. Adding more granular time options hampers the performance of the component. However, the users often need to specify time intervals that does not align with the options present. This change adds the `creatable` flag to the time select to enable this.

Please note that the `creatable` flag in the  underlying BaseWeb component is not quite intuitive.

* fix hasMissingEvents for timeouts

* Create useThrottledState

* use defaultThrottleSettings

* fix failing test

* Show loading indicator and load missing events

* fix lint

* fix types

* fix lint

* revert status changes

* rename helper

* rename method

* remove statusReady from WorkflowHistoryTimelineGroup

---------

Co-authored-by: Pandey Anshuman Kishore <pandeyanshuman@users.noreply.github.com>
Co-authored-by: Adhitya Mamallan <adhitya.mamallan@uber.com>