Fix bug with ungrouped history table when selected event is hidden (#987)
When the ungrouped table view is opened with a selected event that is then hidden by the filters, Virtuoso is unable to find the selected event and so, it crashes. This PR fixes the issue by not doing the check if the selected event is not present in the list of events to show.