feat: Implement optional throttling for fetching history pages in WorkflowHistoryFetcher (#1122)
Implement throttling of calls to fetchNextPage in WorkflowHistoryFetcher's start method, which makes calls to keep fetching workflow history until a condition is met, or there is no more to load. This has been done to prepare for eagerly loading workflow history, to prevent rapid successive calls to fetch history for a running workflow.

Signed-off-by: Adhitya Mamallan <adhitya.mamallan@uber.com>